+++ packages/core/src/cache/index.ts
@@ -1,4 +1,5 @@
 export * from "./constants.js";
+export * from "./empty-populate.js";
 export * from "./errors.js";
 export * from "./fetch.js";
 export * from "./io.js";